+++ src/engine/Vector2D.cpp
@@ -1,6 +1,6 @@
 #include "engine/Vector2D.h"
 
-Vector2D::Vector2D(int x, int y) :
+Vector2D::Vector2D(float x, float y) :
     mX { x },
     mY { y }
     {};
